Улучшила читаемость тестов за счет замены assert на AssertionError

+++ tests/api/auth/test_user_negative.py
@@ -27,7 +27,11 @@ class TestUserAPINegative:
             common_user: User,
             db_helper: DBHelper,
             expected_status: HTTPStatus = HTTPStatus.FORBIDDEN):
-        assert db_helper.user_exists_by_email(common_user.email)
+        # assert db_helper.user_exists_by_email(common_user.email)
+        if not db_helper.user_exists_by_email(common_user.email):
+            raise AssertionError(
+                f"Пользователь {common_user.email} не существует в БД"
+            )
         response = common_user.api.user_api.get_user_info(
             user_id=common_user.id, expected_status=expected_status)
         forbidden_response = UserForbiddenResponse.model_validate(
@@ -62,13 +66,22 @@ class TestUserAPINegative:
             common_user: User,
             db_helper: DBHelper,
             expected_status: HTTPStatus = HTTPStatus.FORBIDDEN):
-        assert db_helper.user_exists_by_email(common_user.email)
+        # assert db_helper.user_exists_by_email(common_user.email)
+        if not db_helper.user_exists_by_email(common_user.email):
+            raise AssertionError(
+                f"Пользователь {common_user.email} не существует в БД"
+            )
         response = regular_user.api.user_api.delete_user(
             common_user.id, expected_status=expected_status)
         forbidden_response = UserForbiddenResponse.model_validate(
             response.json())
         assert forbidden_response.message == "Forbidden"
-        assert db_helper.user_exists_by_email(common_user.email)
+        # assert db_helper.user_exists_by_email(common_user.email)
+        if not db_helper.user_exists_by_email(common_user.email):
+            raise AssertionError(
+                f"Пользователь {common_user.email} не существует в БД, "
+                "ожидалось, что у пользователя нет права удалять пользователей"
+            )
 
     @pytest.mark.smoke
     @pytest.mark.negative
@@ -79,8 +92,13 @@ class TestUserAPINegative:
             super_admin,
             db_helper: DBHelper,
             expected_status: HTTPStatus = HTTPStatus.NOT_FOUND):
-        assert db_helper.get_user_by_id(
-            CommonConstants.NON_EXISTENT_USER_ID) is None
+        # assert db_helper.get_user_by_id(
+        #     CommonConstants.NON_EXISTENT_USER_ID) is None
+        if db_helper.get_user_by_id(CommonConstants.NON_EXISTENT_USER_ID):
+            raise AssertionError(
+                f"Пользователь с ID {CommonConstants.NON_EXISTENT_USER_ID} существует в БД, "
+                "ожидалось, что пользователь с таким ID не существует в БД"
+            )    
         response = super_admin.api.user_api.delete_user(
             CommonConstants.NON_EXISTENT_USER_ID,
             expected_status=expected_status)
